+++ header/light.hpp
@@ -9,7 +9,7 @@ class Light {
 
     public:
         Light(Color i);
-        virtual Color calculate_intensity(Vector P, Vector N, Vector V, int s, Object* o, bool has_shadow) = 0;
+        virtual Color calculate_intensity(Vector P, Vector N, Vector V, double s, Object* o, bool has_shadow) = 0;
         virtual Vector get_l(Vector P) = 0;
         double specular(Vector N, Vector L, Vector V, double s);
