+++ header/point_light.hpp
@@ -9,7 +9,7 @@ class PointLight : public Light {
         
     public:
         PointLight(Color intensity, Vector position);
-        Color calculate_intensity(Vector P, Vector N, Vector V, int s,  Object* o, bool has_shadow);
+        Color calculate_intensity(Vector P, Vector N, Vector V, double s,  Object* o, bool has_shadow);
         Vector get_l(Vector P);
     
         Vector get_position();
